requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
Django==3.0.3
djangorestframework==3.11.0
importlib-metadata==1.5.0
kombu==4.6.7
pytz==2019.3
redis==3.4.1
sqlparse==0.3.0
vine==1.3.0
zipp==2.2.0
